+++ devices/temp_freebsd.go
@@ -21,7 +21,7 @@ var sensorOIDS = map[string]string{
 }
 
 func update(temps map[string]int) map[string]error {
-	var errors map[string]error
+	errors := make(map[string]error)
 
 	for k, v := range sensorOIDS {
 		if _, ok := temps[k]; !ok {